feat: migrate to Qt 6.8 and polish UI scaling/assets
Qt 5.15 -> Qt 6.8 (MSVC 2022):
- find_package(Qt6)/Qt6 targets, CMake >= 3.22, Qt6_DIR
- QMouseEvent::globalPos() -> globalPosition().toPoint()
- enterEvent(QEvent*) -> enterEvent(QEnterEvent*)
- qsizetype narrowing for .size() returns
- OpenSSL runtime DLL names 1.1 -> 3.x

UI scaling & assets:
- scale window sizes proportionally to the screen
- wrap settings dialog in a QScrollArea to avoid small-screen overflow
- copy resources/*.png next to the executable; relative image paths
- scale the pet avatar with the window size (aspect ratio preserved)

Fix multimodal (array-form) content parsing in SSE streaming.

Replace bundled OpenSSL runtime with 3.x and add PNG assets.

+++ src/app/ChatController.cpp
@@ -173,7 +173,7 @@ void ChatController::cancelAll()
 
 int ChatController::pendingRequestCount() const
 {
-    return pending_.size();
+    return static_cast<int>(pending_.size());
 }
 
 AppError ChatController::lastAppError() const
